docs(agr): add Kdoc with @param and @return for functions
