+++ algorithm/src/main/kotlin/io/github/wliamp/agr/impl/CriteriaImpl.kt
@@ -2,7 +2,7 @@ package io.github.wliamp.agr.impl
 
 import reactor.core.publisher.Mono
 
-class EqualsCriteria<T, V>(
+internal class EqualsCriteria<T, V>(
     private val selector: (T) -> V,
     private val expected: V
 ) : ICriteria<T> {
@@ -10,7 +10,7 @@ class EqualsCriteria<T, V>(
         Mono.fromSupplier { selector(target) == expected }
 }
 
-class RangeCriteria<T : Comparable<T>>(
+internal class RangeCriteria<T : Comparable<T>>(
     private val selector: (T) -> T,
     private val min: T,
     private val max: T
@@ -22,7 +22,7 @@ class RangeCriteria<T : Comparable<T>>(
         }
 }
 
-class GreaterThanCriteria<T : Comparable<T>>(
+internal class GreaterThanCriteria<T : Comparable<T>>(
     private val selector: (T) -> T,
     private val threshold: T
 ) : ICriteria<T> {
@@ -30,7 +30,7 @@ class GreaterThanCriteria<T : Comparable<T>>(
         Mono.fromSupplier { selector(target) > threshold }
 }
 
-class LessThanCriteria<T : Comparable<T>>(
+internal class LessThanCriteria<T : Comparable<T>>(
     private val selector: (T) -> T,
     private val threshold: T
 ) : ICriteria<T> {
@@ -38,7 +38,7 @@ class LessThanCriteria<T : Comparable<T>>(
         Mono.fromSupplier { selector(target) < threshold }
 }
 
-class ContainsCriteria<T, E>(
+internal class ContainsCriteria<T, E>(
     private val selector: (T) -> Collection<E>,
     private val expected: E
 ) : ICriteria<T> {
@@ -46,7 +46,7 @@ class ContainsCriteria<T, E>(
         Mono.fromSupplier { selector(target).contains(expected) }
 }
 
-class StartsWithCriteria<T>(
+internal class StartsWithCriteria<T>(
     private val selector: (T) -> String,
     private val prefix: String
 ) : ICriteria<T> {
@@ -54,7 +54,7 @@ class StartsWithCriteria<T>(
         Mono.fromSupplier { selector(target).startsWith(prefix) }
 }
 
-class EndsWithCriteria<T>(
+internal class EndsWithCriteria<T>(
     private val selector: (T) -> String,
     private val suffix: String
 ) : ICriteria<T> {
@@ -62,14 +62,14 @@ class EndsWithCriteria<T>(
         Mono.fromSupplier { selector(target).endsWith(suffix) }
 }
 
-class CustomCriteria<T>(
+internal class CustomCriteria<T>(
     private val predicate: (T) -> Boolean
 ) : ICriteria<T> {
     override fun matches(target: T): Mono<Boolean> =
         Mono.fromSupplier { predicate(target) }
 }
 
-class AndCriteria<T>(
+internal class AndCriteria<T>(
     private val left: ICriteria<T>,
     private val right: ICriteria<T>
 ) : ICriteria<T> {
@@ -78,7 +78,7 @@ class AndCriteria<T>(
             .zipWith(right.matches(target)) { l, r -> l && r }
 }
 
-class OrCriteria<T>(
+internal class OrCriteria<T>(
     private val left: ICriteria<T>,
     private val right: ICriteria<T>
 ) : ICriteria<T> {
@@ -87,7 +87,7 @@ class OrCriteria<T>(
             .zipWith(right.matches(target)) { l, r -> l || r }
 }
 
-class NotCriteria<T>(
+internal class NotCriteria<T>(
     private val inner: ICriteria<T>
 ) : ICriteria<T> {
     override fun matches(target: T): Mono<Boolean> =
